OY-3791 : SchedulerConfig lisätty.
